move bounding box logic to base tooltip
