+++ src/components/Hoverable/index.js
@@ -42,11 +42,6 @@ class Hoverable extends Component {
                 if (!scrolling && this.isHoveredRef) {
                     this.setState({isHovered: this.isHoveredRef}, this.props.onHoverIn);
                 } else if (scrolling && this.isHoveredRef) {
-                    /**
-                     * If the user has started scrolling and the isHoveredRef is true, then we should set the hover state to false.
-                     * This is to hide the existing hover and reaction bar.
-                     */
-                    this.isHoveredRef = false;
                     this.setState({isHovered: false}, this.props.onHoverOut);
                 }
                 this.isScrollingRef = scrolling;
